refactor: changed file structure for project.py
